[TOOL] Add dynamic creation of rules for object file
The analyzing process requires the creation of the obkect file, starting
from the source file. Before, there was a generic rules for all object
files we have to create. Now, starting from the list of all object files
we aspect to generate, a list of rules of each file is dynamically
created.

Note: This will be useful when the external dependencies management will
be added because in that case the whole path of the source file is
different between various file and the root director is not the same
between source file and object file (one for all object file).