[TOOL] Add extenal dependecy management
Now, a project can explicit dependencies to other projects present into
the library.
To add one or more dependencies, the file dependency.mk have to be
present into the project's root folder with the follow populated
variable:

DIPENDENCE_PROJECTS += <list of the name of external project to include>

An example is:

	DIPENDENCE_PROJECTS += \
		clk_2N_divider \
		frequency_counter

Where clk_2N_divider and frequency_counter are two project already
present inside the library.

Note: the absence of this addiction file does not generate errors
because it is an optional file